+++ app/admin/user.rb
@@ -317,7 +317,11 @@ ActiveAdmin.register User do
   config.action_items.delete_at(2)
 
   action_item :destroy, only: :show do
-    link_to t('active_admin.user.delete'), { action: :destroy }, method: :delete, data: { confirm: t('active_admin.user.delete_confirmation', count: 1) }
+    if Match.sent.where(expert: user.single_user_experts).in_progress.exists?
+      button_tag t('active_admin.user.delete'), title: t('activerecord.errors.models.expert.attributes.base.cant_delete_experts_with_in_progress_matches'), disabled: true, class: "button-base"
+    else
+      link_to t('active_admin.user.delete'), { action: :destroy }, method: :delete, data: { confirm: t('active_admin.user.delete_confirmation', count: 1) }
+    end
   end
 
   member_action :normalize_values do
